+++ go/go.mod
@@ -6,7 +6,7 @@ require (
 	github.com/beanstalkd/go-beanstalk v0.2.0
 	github.com/biogo/hts v1.4.4
 	github.com/bytedance/sonic v1.10.2
-	github.com/dustin/go-humanize v1.0.1
+	github.com/klauspost/compress v1.17.3
 	github.com/klauspost/pgzip v1.2.6
 	github.com/opensearch-project/opensearch-go/v2 v2.3.0
 	gopkg.in/yaml.v3 v3.0.1
@@ -15,7 +15,6 @@ require (
 require (
 	github.com/chenzhuoyu/base64x v0.0.0-20230717121745-296ad89f973d // indirect
 	github.com/chenzhuoyu/iasm v0.9.1 // indirect
-	github.com/klauspost/compress v1.17.3 // indirect
 	github.com/klauspost/cpuid/v2 v2.2.6 // indirect
 	github.com/twitchyliquid64/golang-asm v0.15.1 // indirect
 	golang.org/x/arch v0.6.0 // indirect
